+++ app/src/main/java/a/a/a/QB.java
@@ -73,7 +73,7 @@ public class QB extends BaseValidator {
       } 
     } else {
       ArrayList<String> arrayList1 = new ArrayList<>();
-      for (byte b1 = 1; b1 <= this.j; b1++) {
+      for (int b1 = 1; b1 <= this.j; b1++) {
         StringBuilder stringBuilder = new StringBuilder();
         stringBuilder.append(paramString);
         stringBuilder.append("_");
@@ -117,7 +117,7 @@ public class QB extends BaseValidator {
     } 
     String[] arrayOfString = this.f;
     int i = arrayOfString.length;
-    byte b = 0;
+    int b = 0;
     while (true) {
       if (b < i) {
         if (paramString.equals(arrayOfString[b])) {
